+++ src/marais/graphql/dsl/specials.kt
@@ -8,27 +8,33 @@ import kotlin.reflect.KClass
 import kotlin.reflect.KFunction
 import kotlin.reflect.full.primaryConstructor
 
-data class ScalarBuilder(
+class ScalarBuilder(
     val kclass: KClass<*>,
     val name: String,
     val coercing: Coercing<*, *>,
-    val description: String?,
+    description: String?,
     val builder: GraphQLScalarType.Builder.() -> Unit
-)
+) {
+    val description: String? = description ?: kclass.extractDesc()
+}
 
-data class EnumBuilder(
+class EnumBuilder(
     val kclass: KClass<*>,
     val name: String,
-    val description: String?,
+    description: String?,
     val builder: GraphQLEnumType.Builder.() -> Unit
-)
+) {
+    val description: String? = description ?: kclass.extractDesc()
+}
 
-data class InputBuilder(
+class InputBuilder(
     val kclass: KClass<*>,
     val name: String,
-    val description: String?,
+    description: String?,
     val builder: GraphQLInputObjectType.Builder.() -> Unit
 ) {
+    val description: String? = description ?: kclass.extractDesc()
+
     internal val constructor: KFunction<*> =
         kclass.primaryConstructor ?: throw Exception("Can't find a primary constructor for ${kclass.deriveName()}")
     internal val fields = constructor.parameters.map { it.name!! to it.type }